Split `ValidationTest`
Addressing #57.
